FinishedSampleCode/requirements.txt
tornado
dnspython
pymongo
configparser
google-cloud-vision
google-cloud-language
